This is a pretty major divergence from the original django-invitations-simplifed codebase.
Kind of didn't want to do this, but oh well this needs to ship tomorrow. I'll clean it up later.

The requirement that came down the pipe was that when an admin user sends out an invite, they should be able to select from the django.contrib.admin groups that are already defined, and those should get applied to the new user when they accept.

Adding the modelmulitplechoicefield was easy. Getting those choices to go from the form to the invite to the user dropped me into magical syntax failure land. Yes, there are probably better ways to do it than a for loop  that goes through all the group objects one by one, but the couple of other ways I tried had weird behaviors.